refactor(frontend): split playtest sky runtime
